singleDataTransfer with comments

+++ src/singleDataTransfer.c
@@ -2,8 +2,22 @@
 #include "emulator_misc.h"
 #include "armStructure.h"
 
+/* function that executes the single data transfer instructions
+ * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * *
+ * if the flagI is set then Offset is interpreted as a shifted register taking 
+ * into account that a post-indexing load or store in which Rm is the same 
+ * register as Rn is not allowed
+ * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * *
+ * if flagU is set, the offset is added to the base register, otherwise 
+ * is subtracted
+ * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * *
+ * if flagP is set the offset is added/subtracted to the base register before 
+ * transferring the data, otherwise the offset is added/subtracted to the base 
+ * register after transferring.
+ */
 void singleDataTransfer(uint32_t instruction) {
 
+	// if the condition does not hold, the instruction is not executed
 	if(!checkConditionField(instruction)) {
 		return;
 	}
@@ -14,7 +28,7 @@ void singleDataTransfer(uint32_t instruction) {
 	uint32_t flagI = (instruction >> 25) % 2;
 	uint32_t rn = (instruction & MUL_SD_REG_MASK1) >> 16;
 	uint32_t rd = (instruction & MUL_SD_REG_MASK2) >> 12;
-	uint16_t offset = instruction & 0xfff;
+	uint32_t offset = instruction & 0xfff;
 	uint32_t address = ARM.registers[rn];
 	int32_t sign = -1;
 
@@ -26,7 +40,6 @@ void singleDataTransfer(uint32_t instruction) {
 		}
 		offset = DPShift(offset, ((offset >> 1) & 3), instruction);
 	}
-
 
 	if(flagU) {
 		sign = 1;
@@ -42,6 +55,13 @@ void singleDataTransfer(uint32_t instruction) {
 
 }
 
+/* function that load or store 
+ * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * *
+ * if flagU is set, the word is loaded from memory, otherwise the word is stored
+ * into memory
+ * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * *
+ * if the address is out of bounds it will print a message accordingly
+ */
 void load_store(uint32_t rd, uint32_t address, uint32_t flagL) {
 
 	if (flagL) {
@@ -72,9 +92,14 @@ void load_store(uint32_t rd, uint32_t address, uint32_t flagL) {
 		}
 
 	}
+
 }
 
+/* function that prints a message when the address is out of bounds taking into 
+ * consideration particular cases related to pins
+ */
 uint32_t printMessage(uint32_t address) {
+
 	switch(address) {
 		case 0x20200008:
 			printf("One GPIO pin from 20 to 29 has been accessed\n");
@@ -95,4 +120,5 @@ uint32_t printMessage(uint32_t address) {
 			printf("Error: Out of bounds memory access at address 0x%08x\n", address);
 			return 0;
 	}
+
 }
